Move `AccountInputContainer` into `widget` package
